Move resourcemgr into calicoctl and only expose required public methods
